requirements/base.txt
Django==1.8.4
Unipath==1.1
cx-Oracle==5.2
django-discover-runner==1.0
django-role-permissions==0.5.0
django-suit==0.2.14
six==1.9.0